pyproject.toml
[tool.poetry]
name = "rudesktop"
version = "0.1.0"
description = ""
authors = ["Ilya Smelov  <PivnoyFei@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.1.0"
numpy = "^1.26.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
